+++ amqp_message_parser.go
@@ -2,7 +2,7 @@ package gorabbit
 
 import (
 	"errors"
-	"github.com/streadway/amqp"
+	amqp "github.com/rabbitmq/amqp091-go"
 	"strings"
 )