+++ amqp_message_parser_test.go
@@ -2,7 +2,7 @@ package gorabbit
 
 import (
 	"github.com/go-playground/assert/v2"
-	"github.com/streadway/amqp"
+	amqp "github.com/rabbitmq/amqp091-go"
 	"testing"
 )