+++ client_helper.go
@@ -2,7 +2,7 @@ package gorabbit
 
 import (
 	"fmt"
-	"github.com/streadway/amqp"
+	amqp "github.com/rabbitmq/amqp091-go"
 	"time"
 )