+++ model.go
@@ -3,7 +3,7 @@ package gorabbit
 import (
 	"errors"
 	"fmt"
-	"github.com/streadway/amqp"
+	amqp "github.com/rabbitmq/amqp091-go"
 	"time"
 )